+++ wio/Ranger.cpp
@@ -3,7 +3,7 @@
 
 Ranger::Ranger(int pin) : _ultrasonic(pin) { }
 
-long Ranger::MeasureRange()
+double Ranger::MeasureRange()
 {
   return _ultrasonic.MeasureInCentimeters();
 }
